+++ SplitIt/app/src/main/java/project/cs495/splitit/MainActivity.java
@@ -28,11 +28,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
 import com.microblink.EdgeDetectionConfiguration;
 import com.microblink.FrameCharacteristics;
 import com.microblink.IntentUtils;
@@ -45,14 +42,12 @@ import com.scandit.barcodepicker.ScanditLicense;
 
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Iterator;
 import java.util.concurrent.Semaphore;
 
 import project.cs495.splitit.models.Bill;
 import project.cs495.splitit.models.Item;
 import project.cs495.splitit.models.Receipt;
 import project.cs495.splitit.models.ReceiptBuilder;
-import project.cs495.splitit.models.User;
 
 public class MainActivity extends AppCompatActivity
     implements SelectGroupDialogFragment.SelectGroupDialogListener{
@@ -409,128 +404,9 @@ public class MainActivity extends AppCompatActivity
         if (id == R.id.action_settings) {
             accountSettings();
             return true;
-        } else if (id == R.id.sync_bills) {
-            syncBills();
-            return true;
         }
 
         return super.onOptionsItemSelected(item);
-    }
-
-    private void syncBills() {
-        Utils.getDatabaseReference().child("users").addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                for (DataSnapshot userSnapshot : dataSnapshot.getChildren()) {
-                    userSnapshot.child("bills").getRef().removeValue();
-                }
-                parseReceiptsForBillAmounts();
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-    }
-
-    private void parseReceiptsForBillAmounts() {
-        Utils.getDatabaseReference().child("receipts").addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                for (DataSnapshot receiptSnapshot : dataSnapshot.getChildren()) {
-                    Receipt receipt = receiptSnapshot.getValue(Receipt.class);
-                    adjustBillsForSingleReceipt(receipt);
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-    }
-
-    private void adjustBillsForSingleReceipt(Receipt receipt) {
-        final String creatorId = receipt.getCreator();
-        float receiptSubtotal = receipt.getSubtotal();
-        if (receiptSubtotal == 0) {
-            receiptSubtotal = receipt.getPrice();
-        }
-        final float receiptTax = receipt.getTax();
-        final float finalReceiptSubtotal = receiptSubtotal;
-        Utils.getDatabaseReference().child("items").orderByChild("receiptIds/"+receipt.getReceiptId()).equalTo(true).addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                for (DataSnapshot itemSnapshot: dataSnapshot.getChildren()) {
-                    Item item = itemSnapshot.getValue(Item.class);
-                    String assignee = item.getAssignedUser();
-                    if (assignee != null && !creatorId.equals(assignee)) {
-                        float portionOfSubtotal = item.getPrice() / finalReceiptSubtotal;
-                        float itemTax = portionOfSubtotal * receiptTax;
-                        float itemTotalCost = itemTax + item.getPrice();
-                        adjustCreatorAssigneeBillByAmount(creatorId, assignee, itemTotalCost);
-                    }
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-    }
-
-    private void adjustCreatorAssigneeBillByAmount(String creatorId, String assignee, float itemTotalCost) {
-        adjustUserBillByAmount(creatorId, assignee, -1*itemTotalCost);
-        adjustUserBillByAmount(assignee, creatorId, itemTotalCost);
-    }
-
-    private void adjustUserBillByAmount(final String userId, final String billTargetUserId, final float itemTotalCost) {
-            Utils.getDatabaseReference().child("users").child(userId).child("bills").orderByChild("uid").equalTo(billTargetUserId).addListenerForSingleValueEvent(new ValueEventListener() {
-                @Override
-                public void onDataChange(DataSnapshot dataSnapshot) {
-                    Bill bill;
-                    Iterator<DataSnapshot> dataSnapshotIterator = dataSnapshot.getChildren().iterator();
-                    if (dataSnapshotIterator.hasNext()) {
-                        DataSnapshot billSnapshot = dataSnapshotIterator.next();
-                        bill = billSnapshot.getValue(Bill.class);
-                        Float billCurrentAmount = bill.getAmountAsFloat();
-                        bill.setAmountFromFloat(billCurrentAmount + itemTotalCost);
-                        billSnapshot.getRef().child("amount").setValue(bill.getAmount());
-                    } else {
-                        bill = new Bill();
-                        bill.setUid(billTargetUserId);
-                        bill.setAmountFromFloat(itemTotalCost);
-                        String billKey = Utils.getDatabaseReference().child("users").child(userId).child("bills").push().getKey();
-                        Utils.getDatabaseReference().child("users").child(userId).child("bills").child(billKey).setValue(bill);
-                        saveTargetUserDetails(userId, billTargetUserId, itemTotalCost, billKey);
-                    }
-
-                }
-
-                @Override
-                public void onCancelled(DatabaseError databaseError) {
-
-                }
-            });
-    }
-
-    private void saveTargetUserDetails(final String userId, final String billTargetUserId, final float itemTotalCost, final String billKey) {
-        Utils.getDatabaseReference().child("users").child(billTargetUserId).addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                User billTargetUser = dataSnapshot.getValue(User.class);
-                DatabaseReference databaseReference = Utils.getDatabaseReference().child("users").child(userId).child("bills").child(billKey);
-                databaseReference.child("name").setValue(billTargetUser.getName());
-                databaseReference.child("email").setValue(billTargetUser.getEmail());
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
     }
 
     /**
