+++ SplitIt/app/src/main/res/menu/account_settings_menu.xml
@@ -7,8 +7,4 @@
         android:orderInCategory="100"
         android:title="@string/action_settings"
         app:showAsAction="never" />
-    <item
-        android:id="@+id/sync_bills"
-        android:title="Sync Bills"
-        app:showAsAction="never" />
 </menu>
